Code/api_gateway/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
grpcio==1.51.3